Phase 4: polish — comments, types, idioms
Translate Dutch comments to English (codebase is otherwise English):
  - client/anim/FPDeflectAnim.java
  - events/DeflectCombat.java
  - events/DeflectServerTicker.java
  - network/packet/S2CDeflectHit.java
  - client/events/FirstPersonDeflectOverlay.java

ModItems.java:
  - drop the ✅ emoji + path-restating comments per project no-emoji preference
  - mark class final + add private constructor (utility class, never instantiated)

DeflectCombat:
  - drop the safeGetData wrapper (one-liner around getOrNull, no longer
    earning its name) and inline its single call site
  - clarify the class doc + Dutch translations
  - drop a stale `// Snelheden`, `// Re-emit van...`, `// Verdere safety`,
    `// Ping` comment cluster — code is self-explanatory after the rename
  - tighten the stationary-projectile skip from a 3-line block + comment
    to a 1-line `if (current < 0.01) continue;`
  - drop unused NetworkDirection import (unused after sendToPlayer migration)

ShurikenCombat:
  - tickCounter int -> long (overflow protection: ~3.4 years -> never)
  - matching long fields on State for tick comparisons

DashResets:
  - HashMap -> ConcurrentHashMap (LivingHurtEvent/LivingDeathEvent can fire
    from worker threads in some Forge versions; matches DragonbladeCombat's
    concurrent-safe pattern)
  - inner per-attacker maps also ConcurrentHashMap

ShurikenEntity:
  - drop vestigial `flightTicks % 1 == 0` (always true) — the particle
    spawn was never actually throttled

Compile clean. Pure cosmetic / type-correctness pass — no behavior change.

+++ src/main/java/com/example/genji/events/ShurikenCombat.java
@@ -43,15 +43,16 @@ public class ShurikenCombat {
 
     // ===== Runtime state =====
     private static final Map<UUID, State> STATES = new HashMap<>();
-    private static int tickCounter = 0;
+    /** Long to avoid overflow on long-running servers (~292 billion years instead of ~3.4 years). */
+    private static long tickCounter = 0;
 
     private static class State {
         boolean primaryHeld = false;
         boolean secondaryHeld = false;
 
-        int sharedCd = 0;         // global cooldown gating starts
-        int burstShotsLeft = 0;   // shots remaining in current M1 burst
-        int nextBurstTick = 0;    // when to fire the next burst shot
+        long sharedCd = 0;         // global cooldown gating starts
+        int  burstShotsLeft = 0;   // shots remaining in current M1 burst
+        long nextBurstTick = 0;    // when to fire the next burst shot
 
         // For precise 0.68s M2 recovery (13.6 ticks): 14,14,14,13,13 repeat
         int secondaryCycle = 0;
